+++ src/TheLedgerCompany/Models/Balance.cs
@@ -4,8 +4,8 @@ namespace TheLedgerCompany.Models
     {
         public string BankName { get; set; }
         public string BorrowerName { get; set; }
-        public int AmountPaid { get; set; }
-        public int PendingEmis { get; set; }
+        public double AmountPaid { get; set; }
+        public double PendingEmis { get; set; }
 
         public override string ToString()
         {
